Extracted Discord mocking test case to its own class.
